requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30       
cfgv==3.4.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4        
decorator==5.1.1
distlib==0.3.8
Django==5.1.1
django-extensions==3.2.3 
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-jsonmask==0.4.0
drf-yasg==1.21.7
executing==2.1.0
factory_boy==3.3.1
Faker==28.4.1
filelock==3.16.0
identify==2.6.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.27.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonmask==0.1.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nodeenv==1.9.1
openapi-codec==1.3.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.3
pluggy==1.5.0
pre-commit==3.8.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.4
wcwidth==0.2.13